+++ src/constraint_handler/__init__.py
@@ -2,5 +2,5 @@
 The constraint-handler project.
 """
 
-from constraint_handler.main import add_to_control, add_to_globals, set_globals
+from constraint_handler.main import add_declarations, add_to_control, add_to_globals, find_values, set_globals
 from constraint_handler.myClorm import findInControl, findInModel, findInPropagateInit